Create python branch with UI
Needed Libraries

-m pip install pyautogui keyboard

In this example, the aim function takes four arguments (x, y, w, and h) that define the region. These arguments are passed to the locateOnScreen function when it is called in the aim function.

The start_aim function is called when the user clicks the 'Start' button. It reads the values entered in the input fields and registers the 'v' hotkey with the aim function using a lambda function that passes the x, y, w, and h arguments.