Create pixel_color_UI
This is a Python program that uses the PyAutoGUI library to locate certain images on the screen and click on them. The program uses the keyboard module to add hotkeys for starting and stopping the script.

The program has been updated to include a graphical user interface (GUI) built with the tkinter library. The GUI allows the user to input the x, y, width, and height values of the region where the program should search for the image.

The "aim" function is the core of the program, which runs in an infinite loop. Within this loop, it uses the PyAutoGUI library to search for an image called "hp.png" and "target.png" within the specified region on the screen. If the "hp.png" image is found, the program waits for 0.1 seconds before checking again. If the "target.png" image is found, the program uses the PyAutoGUI library to click on the center of the image.

The "kill_switch" function is a callback function that is called when the "q" hotkey is pressed. This function simply exits the program.

The "start_aim" function is called when the user clicks the "Start" button on the GUI. This function gets the x, y, width, and height values from the GUI inputs, adds the "v" and "q" hotkeys, and starts the infinite loop in the "aim" function.

The program has been updated to allow the user to select a pixel color using the tkinter color chooser dialog. The selected pixel color is passed as an argument to the "aim" function, which uses it to search for the "target.png" image on the screen. If the image is found, the program clicks on the center of the image.